[issues-5] Adds helper function for nth-child-hack
In order to support IE8 the adjacent selector is abused to handle
psudo support for nth-child on IE8. It is not full support since
there is a limit to the number of items that can be supported.
However it does give the dynamic nature for the elements that fall
within the limit. For best IE8 support the data-attribute should
be used.