[issues-33] Allows targeting styles when a grid has a specific number of items
There are times when specific layouts work only where there are
greater then a specific number of items. For example if you have
a fluid 3 column grid then having only 2 items makes it look a
bit weird with the hanging space.

This option allows for styles to be generated that will allow for
a fluid 3 column grid but also when the grid only has two items
it will become a two column grid.

This ability is only avaiable when nth-of-type support is enabled.